pyproject.toml
[tool.poetry]
name = "myblog"
version = "0.1.0"
description = "Python-Based (Django) Blog Engine"
authors = ["Grant Ramsay <seapagan@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-ckeditor = "^6.5.1"
django-compressor = "^4.3.1"
django-gravatar2 = "^1.4.4"
django-htmlmin = "^0.11.0"
django-maintenance-mode = "^0.18.0"
django-recaptcha = "^3.0.0"
django-rundevserver = "^0.3.1"
django-secretballot = "^2.0.0"
django-xforwardedfor-middleware = "^2.0"
djangorestframework = "^3.14.0"
dj-hitcount = "^1.3.0"
geoip2 = "^4.6.0"
pillow = "^9.5.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
requests = "^2.29.0"
gunicorn = "^20.1.0"
django-likes = { git = "https://github.com/seapagan/django-likes.git", rev = "develop" }
dj-pagination = { git = "https://github.com/seapagan/dj-pagination.git", rev = "master" }
django-user-sessions = "^2.0.0"


[tool.poetry.group.dev.dependencies]
newrelic = "^8.8.0"
pytest = "^7.3.1"
django-pytest = "^0.2.0"
pytest-randomly = "^3.12.0"
pytest-reverse = "^1.5.0"
pytest-xdist = "^3.2.1"
pytest-cov = "^4.0.0"
pytest-sugar = "^0.9.7"
ipython = "^8.12.0"
pylint = "^2.17.3"
black = "^23.3.0"
flake8 = "^6.0.0"
pydocstyle = "^6.3.0"
flake8-docstrings = "^1.7.0"
pep8-naming = "^0.13.3"
poethepoet = "^0.20.0"
pre-commit = "^3.3.1"
flake8-pyproject = "^1.2.3"
bandit = "^1.7.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
pre = "pre-commit run --all-files"
serve = "python manage.py runserver"

[tool.bandit]
exclude_dirs = ["config/settings.py", "schemas/examples.py"]

[tool.bandit.assert_used]
skips = ['*_test.py', '*/test_*.py']

[tool.pydocstyle]
add-ignore = ["D104"]

[tool.pytest.ini_options]
filterwarnings = [
  "ignore::django.utils.deprecation.RemovedInDjango40Warning",
  "ignore::django.utils.deprecation.RemovedInDjango41Warning",
]

[tool.flake8]
exclude = ["__init__.py", ".git", "*/migrations/*.py", "plugins", "settings.py"]
max-line-length = 80
docstring-convention = "google"
